move _standardize_struct and _prep_augmented_structure_for_bonding to helpers.py
